# internal/api/docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: VoHive 主服务控制面 API — 自动生成 OpenAPI 文档 POC
  title: VoHive Main API
  version: "1.0"
paths:
...
  /devices/{device_id}/operator_selection:
    get:
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceMgmtGetOperatorSelection
      tags:
      - devices
    post:
      consumes:
      - application/json
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceMgmtSetOperatorSelection
      tags:
      - devices
  /devices/{device_id}/operator_selection/scan:
    get:
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceMgmtOperatorScan
      tags:
      - devices
  /devices/{device_id}/operator_selection/scan/stream:
    get:
      parameters:
      - description: device_id
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: DeviceMgmtOperatorScanStream
      tags:
      - devices
...
  /system/pcsc-driver:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: GetPcscDriverStatus
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"